back/ms-transcribe/requirements.txt
absl-py==2.0.0
accelerate
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
awscli==1.29.61
Babel==2.13.0
bangla==0.0.2
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
boto3==1.28.61
botocore==1.31.61
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid
click==8.1.7
clip-anytorch
colorama==0.4.4
contourpy==1.1.1
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.30
dateparser==1.1.8
dctorch
decorator==4.4.2
dnspython==2.4.2
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
dtw-python==1.3.0
einops==0.6.1
encodec
fastapi==0.103.1
ffmpeg-python==0.2.0
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
g2pkk==0.1.2
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gTTS==2.4.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
inflect==5.6.2
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
k-diffusion
kiwisolver==1.4.5
kornia
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
mercurial==6.5.1
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mutagen==1.47.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
openai-whisper
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
proglog==0.1.10
protobuf==4.24.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pymongo==4.5.0
pynndescent==0.5.10
pyparsing==3.1.1
PyPDF2==3.0.1
pypinyin==0.49.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pyttsx3==2.90
pytz==2023.3.post1
pywatchman==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rotary-embedding-torch
rpds-py==0.10.4
rsa==4.7.2
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.2
semantic-version==2.10.0
sentry-sdk==1.31.0
setproctitle==1.3.3
setuptools-rust==1.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.6
starlette==0.27.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.3.1
tokenizers==0.13.3
torch==2.0.1
torchaudio
torchdiffeq
torchsde
torchvision
tqdm==4.64.1
trainer==0.0.31
trampoline==0.1.2
transformers==4.33.3
TTS==0.17.7
typing_extensions==4.7.1
tzlocal==5.1
umap-learn==0.5.4
Unidecode==1.3.7
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.17.0
wandb==0.15.12
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
Werkzeug==3.0.0
whisper-timestamped @ git+https://github.com/linto-ai/whisper-timestamped@127e8b738756420487210490d09938593244b3dd
yarl==1.9.2
